Добавил класс DepartmentService
